renamed start
renamed start to startpage so its more human readable, start could mean anything
