+++ src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import './App.css';
 import LogInPage from './webpages/loginpage/loginpage';
 import SignUpPage from './webpages/signuppage/signuppage';
-import StartPage from './webpages/Start/startpage'; // Import the Start component
+import StartPage from './webpages/startpage/startpage'; // Import the Start component
 import Homepage from './webpages/homepage/homepage';
 import AccountSettings from './webpages/accountsettings/accountsettings';
